CT Thursday AM Push

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Accounts.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public class Accounts {
 	
-	private int accountId;
-	private int userId;
+	private long accountId;
+	private long userId;
 	private BigDecimal balance;
 	
 	public Accounts() {
@@ -21,16 +21,16 @@ public class Accounts {
 		this.balance = balance;
 	}
 
-	public int getAccountId() {
+	public long getAccountId() {
 		return accountId;
 	}
-	public void setAccountId(int accountId) {
+	public void setAccountId(long accountId) {
 		this.accountId = accountId;
 	}
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
 	public BigDecimal getBalance() {
